chore: rename DTO to Dto
