+++ src/Collections/FieldCollection.php
@@ -6,7 +6,7 @@ namespace Innobrain\Structure\Collections;
 
 use Illuminate\Support\Collection;
 use Innobrain\Structure\Builders\FieldFilterBuilder;
-use Innobrain\Structure\DTOs\Field;
+use Innobrain\Structure\Dtos\Field;
 
 /**
  * @extends Collection<string, Field>
