+++ src/Converters/Concerns/ConvertStrategy.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Innobrain\Structure\Converters\Concerns;
 
-use Innobrain\Structure\DTOs\Field;
-use Innobrain\Structure\DTOs\FieldDependency;
-use Innobrain\Structure\DTOs\FieldFilter;
-use Innobrain\Structure\DTOs\Module;
-use Innobrain\Structure\DTOs\PermittedValue;
+use Innobrain\Structure\Dtos\Field;
+use Innobrain\Structure\Dtos\FieldDependency;
+use Innobrain\Structure\Dtos\FieldFilter;
+use Innobrain\Structure\Dtos\Module;
+use Innobrain\Structure\Dtos\PermittedValue;
 
 /**
  * Type-safe converter strategy contract – one method per DTO.
