+++ src/Converters/LaravelRules/LaravelRulesFieldDependency.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Innobrain\Structure\Converters\LaravelRules;
 
-use Innobrain\Structure\DTOs\FieldDependency;
+use Innobrain\Structure\Dtos\FieldDependency;
 
 trait LaravelRulesFieldDependency
 {
