+++ src/Converters/LaravelRules/LaravelRulesModule.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Innobrain\Structure\Converters\LaravelRules;
 
-use Innobrain\Structure\DTOs\Field;
-use Innobrain\Structure\DTOs\Module;
+use Innobrain\Structure\Dtos\Field;
+use Innobrain\Structure\Dtos\Module;
 use Innobrain\Structure\Enums\FieldType;
 
 trait LaravelRulesModule
