+++ src/Dtos/PermittedValue.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Innobrain\Structure\DTOs;
+namespace Innobrain\Structure\Dtos;
 
 use Innobrain\Structure\Concerns\HasConverter;
 use Innobrain\Structure\Contracts\Convertible;
 
-class FieldDependency implements Convertible
+class PermittedValue implements Convertible
 {
     use HasConverter;
 
     public function __construct(
-        public readonly string $dependentFieldKey,
-        public readonly string $dependentFieldValue,
+        public readonly string $key,
+        public readonly string $label,
     ) {}
 }
